requirements.txt
google-genai
mcp-flight-search
mcp
aiohttp
google-generativeai 
crewai 
crewai-tools